backend/requirements.txt
Django==3.1.7

django-cors-headers==3.7.0
djangorestframework==3.12.4
djoser==2.1.0
social-auth-app-django==4.0.0

Pillow==8.1.2

psycopg2==2.8.6

argon2-cffi==20.1.0
django-environ==0.4.5

braintree==4.7.0
stripe==2.58.0